+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support for the latest 2D-Doc specification 3.3.8 document types.
 - HTTP headers when fetching certificate chains to avoid `403` errors.
+- Update to latest TSL list and verify signature when fetching certificates.
 
 ### Changed
 - Fields `18`/`19`/`1A`/`1B` are now typed as alphanumeric (`StringAZ09Sp`).